+++ meta_package_manager/cli.py
@@ -47,7 +47,12 @@ from click_extra import (
 )
 from click_extra.colorize import KO, OK, default_theme as theme, highlight
 from click_extra.commands import default_extra_params
-from click_extra.table import SERIALIZATION_FORMATS, TableFormat, TableFormatOption
+from click_extra.table import (
+    SERIALIZATION_FORMATS,
+    TableFormat,
+    TableFormatOption,
+    print_data,
+)
 from extra_platforms import reduce
 
 from . import __version__, bar_plugin
@@ -63,7 +68,6 @@ from .output import (
     BarPluginRenderer,
     SortableField,
     SortedTableFormatOption,
-    print_serialized,
     print_stats,
 )
 from .pool import pool
@@ -533,7 +537,7 @@ def managers(ctx):
                 {expt.error for expt in manager.cli_errors},
             )
 
-        print_serialized(manager_data, table_format)
+        print_data(manager_data, table_format, root_element="mpm", package="meta-package-manager")
         ctx.exit()
 
     # Human-friendly content rendering.
@@ -660,7 +664,7 @@ def installed(ctx, duplicates):
     # Machine-friendly data rendering.
     table_format = ctx.meta["click_extra.table_format"]
     if table_format in SERIALIZATION_FORMATS:
-        print_serialized(installed_data, table_format)
+        print_data(installed_data, table_format, root_element="mpm", package="meta-package-manager")
         ctx.exit()
 
     # Human-friendly content rendering.
@@ -745,7 +749,7 @@ def outdated(ctx, plugin_output):
     # Machine-friendly data rendering.
     table_format = ctx.meta["click_extra.table_format"]
     if table_format in SERIALIZATION_FORMATS:
-        print_serialized(outdated_data, table_format)
+        print_data(outdated_data, table_format, root_element="mpm", package="meta-package-manager")
         ctx.exit()
 
     # Xbar/SwiftBar-friendly plugin rendering.
@@ -864,7 +868,7 @@ def search(ctx, extended, exact, refilter, query):
     # Machine-friendly data rendering.
     table_format = ctx.meta["click_extra.table_format"]
     if table_format in SERIALIZATION_FORMATS:
-        print_serialized(matches, table_format)
+        print_data(matches, table_format, root_element="mpm", package="meta-package-manager")
         ctx.exit()
 
     # Prepare highlighting helpers.
@@ -939,7 +943,7 @@ def which(ctx, cli_names):
             }
             for manager in ctx.obj.selected_managers()
         ]
-        print_serialized(cli_data, table_format)
+        print_data(cli_data, table_format, root_element="mpm", package="meta-package-manager")
         ctx.exit()
 
     # Print table.